refactor(cache): move code to `dfx-core` (#3036)
