+++ Cargo.toml
@@ -58,6 +58,7 @@ slog = "2.5.2"
 slog-async = "2.4.0"
 slog-term = "2.9.0"
 tar = "0.4.38"
+semver = "1.0.6"
 tempfile = "3.3.0"
 thiserror = "1.0.24"
 time = "0.3.9"
