+++ src/dfx-core/Cargo.toml
@@ -28,6 +28,7 @@ lazy_static.workspace = true
 ring.workspace = true
 schemars.workspace = true
 sec1 = { workspace = true, features = ["std"] }
+semver.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 slog = { workspace = true, features = ["max_level_trace"] }
