+++ src/dfx-core/src/config/mod.rs
@@ -1,2 +1,3 @@
+pub mod cache;
 pub mod directories;
 pub mod model;
