+++ src/dfx/src/commands/cache/install.rs
@@ -1,3 +1,4 @@
+use crate::config::cache::DiskBasedCache;
 use crate::lib::environment::Environment;
 use crate::lib::error::{DfxError, DfxResult};
 
@@ -9,5 +10,6 @@ use clap::Parser;
 pub struct CacheInstall {}
 
 pub fn exec(env: &dyn Environment, _opts: CacheInstall) -> DfxResult {
-    env.get_cache().force_install().map_err(DfxError::from)
+    DiskBasedCache::force_install(&env.get_cache().version_str()).map_err(DfxError::from)?;
+    Ok(())
 }
