+++ src/dfx/src/commands/cache/show.rs
@@ -1,4 +1,3 @@
-use crate::config::cache;
 use crate::lib::environment::Environment;
 use crate::lib::error::DfxResult;
 
@@ -11,6 +10,11 @@ pub struct CacheShowOpts {}
 
 pub fn exec(env: &dyn Environment, _opts: CacheShowOpts) -> DfxResult {
     let v = format!("{}", env.get_version());
-    println!("{}", cache::get_bin_cache(&v)?.as_path().display());
+    println!(
+        "{}",
+        dfx_core::config::cache::get_bin_cache(&v)?
+            .as_path()
+            .display()
+    );
     Ok(())
 }
