+++ src/dfx/src/commands/pull.rs
@@ -1,4 +1,3 @@
-use crate::config::cache::get_cache_root;
 use crate::lib::environment::AgentEnvironment;
 use crate::lib::error::DfxResult;
 use crate::lib::identity::identity_utils::CallSender;
@@ -7,6 +6,7 @@ use crate::lib::operations::canister::get_canister_status;
 use crate::lib::root_key::fetch_root_key_if_needed;
 use crate::lib::{agent::create_agent_environment, environment::Environment};
 use crate::NetworkOpt;
+use dfx_core::config::cache::get_cache_root;
 use dfx_core::config::model::dfinity::CanisterTypeProperties;
 use std::collections::{BTreeMap, BTreeSet, VecDeque};
 use std::io::Write;
