+++ src/dfx/src/lib/error/extension.rs
@@ -6,7 +6,7 @@ use thiserror::Error;
 pub enum ExtensionError {
     // errors related to extension directory management
     #[error("Cannot find cache directory '{0}': {1}")]
-    FindCacheDirectoryFailed(std::path::PathBuf, crate::lib::error::CacheError),
+    FindCacheDirectoryFailed(std::path::PathBuf, dfx_core::error::cache::CacheError),
 
     #[error("Cannot get extensions directory: {0}")]
     EnsureExtensionDirExistsFailed(dfx_core::error::io::IoError),
