+++ src/dfx/src/lib/error/mod.rs
@@ -1,10 +1,8 @@
 pub mod build;
-pub mod cache;
 pub mod extension;
 pub mod project;
 
 pub use build::BuildError;
-pub use cache::CacheError;
 pub use dfx_core::error::identity::IdentityError;
 pub use extension::ExtensionError;
 pub use project::ProjectError;
