+++ src/dfx/src/main.rs
@@ -93,7 +93,7 @@ fn maybe_redirect_dfx(version: &Version) -> Option<()> {
             );
         }
 
-        match crate::config::cache::call_cached_dfx(version) {
+        match dfx_core::config::cache::call_cached_dfx(version) {
             Ok(status) => std::process::exit(status.code().unwrap_or(0)),
             Err(e) => {
                 eprintln!("Error when trying to forward to project dfx:\n{:?}", e);
